+++ public/components/agents/fim/main.tsx
@@ -4,13 +4,14 @@ import '../../common/modules/module.scss';
 import { connect } from 'react-redux';
 import { PromptNoActiveAgent, PromptNoSelectedAgent } from '../prompts';
 import { compose } from 'redux';
-import { withGuard, withUserAuthorizationPrompt } from '../../common/hocs';
+import { withAgentSupportModule, withGuard, withUserAuthorizationPrompt } from '../../common/hocs';
 
 const mapStateToProps = (state) => ({
   currentAgentData: state.appStateReducers.currentAgentData,
 });
 
 export const MainFim = compose(
+  withAgentSupportModule,
   connect(mapStateToProps),
   withGuard(
     (props) => !(props.currentAgentData && props.currentAgentData.id && props.agent),