+++ public/components/overview/office-panel/config/drilldown-operations-config.tsx
@@ -1,6 +1,5 @@
-'use strict';
 /*
- * Wazuh app - Office 365 Drilldown IP field Config.
+ * Wazuh app - Office 365 Drilldown Operations field Config.
  *
  * Copyright (C) 2015-2021 Wazuh, Inc.
  *
@@ -17,31 +16,21 @@ import { VisCard } from '../../../common/modules/panel';
 import { EuiFlexItem, EuiPanel } from '@elastic/eui';
 import { SecurityAlerts } from '../../../visualize/components';
 
-export const drilldownIPConfig = {
+export const drilldownOperationsConfig = {
   rows: [
     {
       height: 400,
       columns: [
         {
-          width: 30,
-          component: (props) => (
-            <VisCard id="Wazuh-App-Overview-Office-Metric-Stats" tab="office" {...props} />
-          ),
-        },
-        {
-          width: 30,
-          component: (props) => (
-            <VisCard id="Wazuh-App-Overview-Office-Top-Events-Pie" tab="office" {...props} />
-          ),
-        },
-        {
           width: 40,
           component: (props) => (
-            <VisCard
-              id={'Wazuh-App-Overview-Office-User-Operation-Level-Table'}
-              tab="office"
-              {...props}
-            />
+            <VisCard id="Wazuh-App-Overview-Office-Top-Users" tab="office" {...props} />
+          ),
+        },
+        {
+          width: 60,
+          component: (props) => (
+            <VisCard id="Wazuh-App-Overview-Office-Country-Tag-Cloud" tab="office" {...props} />
           ),
         },
       ],
@@ -53,7 +42,7 @@ export const drilldownIPConfig = {
           width: 100,
           component: (props) => (
             <VisCard
-              id="Wazuh-App-Overview-Office-Alerts-Evolution-By-User"
+              id="Wazuh-App-Overview-Office-Alerts-Evolution-By-UserID"
               tab="office"
               {...props}
             />
@@ -73,8 +62,8 @@ export const drilldownIPConfig = {
                     { field: 'icon' },
                     { field: 'timestamp' },
                     { field: 'data.office365.UserId', label: 'User ID' },
+                    { field: 'data.office365.ClientIP', label: 'Client IP' },
                     { field: 'rule.description', label: 'Description' },
-                    { field: 'data.office365.Operation', label: 'Operation' },
                   ]}
                   useAgentColumns={false}
                 />