+++ public/components/overview/office-panel/config/drilldown-rules-config.tsx
@@ -1,6 +1,5 @@
-'use strict';
 /*
- * Wazuh app - Office 365 Drilldown IP field Config.
+ * Wazuh app - Office 365 Drilldown Rules field Config.
  *
  * Copyright (C) 2015-2021 Wazuh, Inc.
  *
@@ -17,7 +16,7 @@ import { VisCard } from '../../../common/modules/panel';
 import { EuiFlexItem, EuiPanel } from '@elastic/eui';
 import { SecurityAlerts } from '../../../visualize/components';
 
-export const drilldownIPConfig = {
+export const drilldownRulesConfig = {
   rows: [
     {
       height: 400,
@@ -25,23 +24,19 @@ export const drilldownIPConfig = {
         {
           width: 30,
           component: (props) => (
-            <VisCard id="Wazuh-App-Overview-Office-Metric-Stats" tab="office" {...props} />
+            <VisCard id="Wazuh-App-Overview-Office-Top-Operations" tab="office" {...props} />
           ),
         },
         {
           width: 30,
           component: (props) => (
-            <VisCard id="Wazuh-App-Overview-Office-Top-Events-Pie" tab="office" {...props} />
+            <VisCard id="Wazuh-App-Overview-Office-Top-Users" tab="office" {...props} />
           ),
         },
         {
           width: 40,
           component: (props) => (
-            <VisCard
-              id={'Wazuh-App-Overview-Office-User-Operation-Level-Table'}
-              tab="office"
-              {...props}
-            />
+            <VisCard id="Wazuh-App-Overview-Office-Country-Tag-Cloud" tab="office" {...props} />
           ),
         },
       ],
@@ -53,7 +48,7 @@ export const drilldownIPConfig = {
           width: 100,
           component: (props) => (
             <VisCard
-              id="Wazuh-App-Overview-Office-Alerts-Evolution-By-User"
+              id="Wazuh-App-Overview-Office-Alerts-Evolution-By-UserID"
               tab="office"
               {...props}
             />
@@ -73,7 +68,7 @@ export const drilldownIPConfig = {
                     { field: 'icon' },
                     { field: 'timestamp' },
                     { field: 'data.office365.UserId', label: 'User ID' },
-                    { field: 'rule.description', label: 'Description' },
+                    { field: 'data.office365.ClientIP', label: 'Client IP' },
                     { field: 'data.office365.Operation', label: 'Operation' },
                   ]}
                   useAgentColumns={false}